Enabled restoring deleted invoices and payments

+++ app/Http/Controllers/PaymentController.php
@@ -71,7 +71,7 @@ class PaymentController extends BaseController
         return $table->addColumn('amount', function ($model) { return Utils::formatMoney($model->amount, $model->currency_id); })
             ->addColumn('payment_date', function ($model) { return Utils::dateToString($model->payment_date); })
             ->addColumn('dropdown', function ($model) {
-                if ($model->is_deleted || $model->invoice_is_deleted) {
+                if ($model->invoice_is_deleted) {
                     return '<div style="height:38px"/>';
                 }
 
@@ -89,6 +89,10 @@ class PaymentController extends BaseController
                     $str .= '<li><a href="javascript:restoreEntity('.$model->public_id.')">'.trans('texts.restore_payment').'</a></li>';
                 }
 
+                if ($model->is_deleted) {
+                    return $str;
+                }
+                
                 return $str.'<li><a href="javascript:deleteEntity('.$model->public_id.')">'.trans('texts.delete_payment').'</a></li></ul>
                         </div>';
             })
